server/requirements.txt
google_api_python_client==1.6.4
google_auth_httplib2==0.0.2
oauth2client==4.1.2
oauthlib==2.0.6
pycrypto==2.6.1
python-dateutil==2.5.3
requests==2.22.0
requests-toolbelt==0.7.0
PyJWT==1.6.4
PyYAML==3.11
validate-email==1.3
validators==0.12.0
